+++ src/Model/Response/Item/ProjectResponseItem.php
@@ -1,0 +1,149 @@
+<?php
+
+namespace KassaCom\SDK\Model\Response\Item;
+
+use KassaCom\SDK\Model\Response\AbstractResponse;
+use KassaCom\SDK\Model\Traits\RecursiveRestoreTrait;
+
+class ProjectResponseItem extends AbstractResponse
+{
+    use RecursiveRestoreTrait;
+
+    /**
+     * @var int
+     */
+    protected $id;
+
+    /**
+     * @var string
+     */
+    protected $name;
+
+    /**
+     * @var string
+     */
+    protected $url;
+
+    /**
+     * @var string[]
+     */
+    protected $paymentMethods;
+
+    /**
+     * @var string|null
+     */
+    protected $logoUrl;
+
+    /**
+     * @return int
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * @param int $id
+     */
+    public function setId($id)
+    {
+        $this->id = $id;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getName()
+    {
+        return $this->name;
+    }
+
+    /**
+     * @param string $name
+     */
+    public function setName($name)
+    {
+        $this->name = $name;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getUrl()
+    {
+        return $this->url;
+    }
+
+    /**
+     * @param string $url
+     */
+    public function setUrl($url)
+    {
+        $this->url = $url;
+
+        return $this;
+    }
+
+    /**
+     * @return string[]
+     */
+    public function getPaymentMethods()
+    {
+        return $this->paymentMethods;
+    }
+
+    /**
+     * @param string[] $paymentMethods
+     */
+    public function setPaymentMethods($paymentMethods)
+    {
+        $this->paymentMethods = $paymentMethods;
+
+        return $this;
+    }
+
+    /**
+     * @return string|null
+     */
+    public function getLogoUrl()
+    {
+        return $this->logoUrl;
+    }
+
+    /**
+     * @param string|null $logoUrl
+     */
+    public function setLogoUrl($logoUrl)
+    {
+        $this->logoUrl = $logoUrl;
+
+        return $this;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getRequiredFields()
+    {
+        return [
+            'id' => self::TYPE_INTEGER,
+            'name' => self::TYPE_STRING,
+            'url' => self::TYPE_STRING,
+            'payment_methods' => [self::TYPE_STRING],
+        ];
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getOptionalFields()
+    {
+        return [
+            'logo_url' => self::TYPE_STRING,
+        ];
+    }
+}